Add parsing notebook with python json module
The python json module is used as a script. This does NOT test the code
included in the notebook in any way. It just checks whether the notebook
(which is really just a json file) has a proper structure. If it doesn't
then the jupyter server would refuse to spawn the notebook